Re-organize Touché dataset paths
