+++ ir_datasets/datasets/__init__.py
@@ -35,7 +35,8 @@ from . import nfcorpus
 from . import natural_questions
 from . import nyt
 from . import pmc
-from . import touche
+from . import touche_image
+from . import touche  # must be after argsme,clueweb12,touche_image
 from . import trec_arabic
 from . import trec_mandarin
 from . import trec_spanish
