+++ ir_datasets/formats/__init__.py
@@ -3,7 +3,8 @@ from .base import BaseDocs, BaseQueries, BaseQrels, BaseScoredDocs, BaseDocPairs
 from .csv_fmt import CsvDocs, CsvQueries, CsvDocPairs
 from .tsv import TsvDocs, TsvQueries, TsvDocPairs
 from .trec import TrecDocs, TrecQueries, TrecXmlQueries, TrecColonQueries, TrecQrels, TrecSubQrels, TrecPrels, TrecScoredDocs, TrecDoc, TitleUrlTextDoc, TrecQuery, TrecSubtopic, TrecQrel, TrecSubQrel, TrecPrel, TrecParsedDoc
-from .touche import ToucheQuery, ToucheTitleQuery, ToucheComparativeQuery, ToucheQualityQrel, ToucheQualityCoherenceQrel, ToucheComparativeStance, ToucheQualityComparativeStanceQrel, ToucheControversialStance, ToucheControversialStanceQrel, TouchePassageDoc, ToucheImageRanking, ToucheImageNode, ToucheImagePage, ToucheImageDoc, ToucheQueries, ToucheTitleQueries, ToucheComparativeQueries, ToucheQrels, ToucheQualityQrels, ToucheQualityCoherenceQrels, ToucheQualityComparativeStanceQrels, ToucheControversialStanceQrels, TouchePassageDocs, ToucheImageDocs
+from .touche import ToucheQuery, ToucheTitleQuery, ToucheComparativeQuery, ToucheQualityQrel, ToucheQualityCoherenceQrel, ToucheComparativeStance, ToucheQualityComparativeStanceQrel, ToucheControversialStance, ToucheControversialStanceQrel, TouchePassageDoc, ToucheQueries, ToucheTitleQueries, ToucheComparativeQueries, ToucheQrels, ToucheQualityQrels, ToucheQualityCoherenceQrels, ToucheQualityComparativeStanceQrels, ToucheControversialStanceQrels, TouchePassageDocs
+from .touche_image import ToucheImageRanking, ToucheImageNode, ToucheImagePage, ToucheImageDoc, ToucheImageDocs
 from .webarc import WarcDocs, WarcDoc
 from .ntcir import NtcirQrels
 from .clirmatrix import CLIRMatrixQueries, CLIRMatrixQrels
